requirements.txt
bestconfig==1.3.5
pyaes==1.6.1
pyasn1==0.4.8
PyYAML==6.0
rsa==4.7.2
Telethon==1.23.0
Wand==0.6.7
aiohttp==3.7.4.post0
python-magic==0.4.24